# cohttp.headers: faster faster comparison