--- a/http/bench/bench.ml
+++ b/http/bench/bench.ml
@@ -2,26 +2,32 @@ module Command = Core.Command
 module Staged = Core.Staged
 open Core_bench
 
-let header_mem =
-  let variations =
-    [
-      [ ("foo", "bar") ];
-      [ ("transfer-encoding", "bar") ];
-      [ ("traNsfer-eNcoding", "bar") ];
-      [ ("traNsfer eNcoding", "bar") ];
-      [ ("traNsfer-eNcodkng", "bar") ];
-      [ ("transfer-encodinG", "bar") ];
-    ]
-  in
-  let key = "transfer-encoding" in
-  List.mapi
-    (fun i header ->
-      let header = Http.Header.of_list header in
-      Bench.Test.create
-        ~name:("Header.mem - " ^ string_of_int i)
-        (fun () ->
-          let (_ : bool) = Http.Header.mem header key in
-          ()))
-    variations
+let header_names = [
+  "Accept";
+  "Accept-Encoding";
+  "Accept-Language";
+  "Cache-Control";
+  "Connection";
+  "Host";
+  "If-Modified-Since";
+  "If-None-Match";
+  "Origin";
+  "Referer";
+  "Sec-Fetch-Dest";
+  "Sec-Fetch-Mode";
+  "Sec-Fetch-Site";
+]
 
-let () = Command.run @@ Bench.make_command header_mem
+let header =
+  header_names
+  |> List.map (fun s -> (s, "value"))
+  |> Http.Header.of_list
+
+let bench_header_mem =
+  Bench.Test.create ~name:"Header.mem"
+    (fun () ->
+       List.iter (fun key -> assert (Http.Header.mem header key))
+         header_names
+    )
+
+let () = Command.run @@ Bench.make_command [ bench_header_mem ]