--- a/http/bench/dune
+++ b/http/bench/dune
@@ -8,4 +8,4 @@
  (enabled_if
   (= %{ocaml-config:word_size} 64))
  (action
-  (run ./bench.exe -quota 5s)))
+  (run ./bench.exe time cycles)))