--- a/http/bench/dune
+++ b/http/bench/dune
@@ -2,10 +2,10 @@
  (name bench)
  (libraries http core core_bench))
 
-(rule
- (alias runtest)
- (package http)
- (enabled_if
-  (= %{ocaml-config:word_size} 64))
- (action
-  (run ./bench.exe time cycles)))
+; (rule
+;  (alias runtest)
+;  (package http)
+;  (enabled_if
+;   (= %{ocaml-config:word_size} 64))
+;  (action
+;   (run ./bench.exe time cycles)))