see
  https://github.com/mirage/ocaml-cohttp/pull/778#issuecomment-1006459031