@gasche 2022-01-06
I'm unfamiliar with cohttp development, so here are some basic questions:

- How does one benchmark this change? @mseri, would you by chance be willing to do the benchmarking? (At least giving me the command to run to benchmark, and what measure to look at, would be helpful.)
- Where do I write tests for this function? There is a risk for off-by-one error / boundary conditions here, that could be solved by just checking the behavior exhaustively for all small lengths. But I did not find where unit tests are located.