@rgrinberg 2022-01-06
> Where do I write tests for this function? There is a risk for off-by-one error / boundary conditions here, that could be solved by just checking the behavior exhaustively for all small lengths. But I did not find where unit tests are located.

The `http` package is very new and the existing tests are in the `cohttp` package. You can add the first test to `http/test` though. To expose the function for unit testing, add a `Private` module either in `Http` or in `Http.Header`.