@gasche 2022-01-06
The CI says that formatting is wrong, and that the String primitives are not exposed on older OCaml versions (4.08-4.12). Both can be easily fixed (Bytes has the primitives exposed), but I would welcome performance feedback first.