@rgrinberg 2022-01-07
I added a very simple benchmark using core_bench that isn't at all realistic but can be easily expanded to include more cases.