@gasche 2022-01-07
Thanks!

Would you by chance be willing to push a similar skeleton for tests, in addition to benchmarks? I don't have your favorite dune-based testing infrastructure in muscle memory, so it would be a bit harder for me. (I hoped to avoid this work by letting a more familiar contributor make a PR :-)

Note (unrelated): the new version only significantly improves over the previous version when the first difference in the string is far in the two strings; in particular, in the common case where the two strings are in fact equal. (When they are  distinct, in the common case their length will differ, so they will be discarded immediately with both the new and old versions.)