@mseri 2022-01-07
Very cool, thanks a lot.

Fwiw the headers module also have fuzzy tests (which we can probably move to the http package in a later PR), so that should help finding out issues in this function.

In the past I did something on the line of what @rgrinberg added, but local to my machine, and I used a more global micro benchmark like the one I added to [multicore httpaf benchmarks](https://github.com/ocaml-multicore/retro-httpaf-bench/blob/master/cohttp-lwt-unix/cohttp_lwt_unix.ml) to check if it affected the overall speed in any way. Nothing scientific in practice.