@mseri 2022-01-07
> Would you by chance be willing to push a similar skeleton for tests, in addition to benchmarks? I don't have your favorite dune-based testing infrastructure in muscle memory, so it would be a bit harder for me. (I hoped to avoid this work by letting a more familiar contributor make a PR :-)

In 1 hour or so I should be able to spend some time to move the headers tests here