@mseri 2022-01-07
I don't have wrk2 available to try the other tests, but your microbenchmark seems quite promising