@mseri 2022-01-07
The Ocaml-CI is actually successfully testing everything, but fails since it tries to build the benchmark anyway.

The github CI, which was succeeding until yesterday, now fails to run the js tests for some reason 🤷