@gasche 2022-01-07
Shouldn't `core_bench` be added as a `test` dependency in `http.opam`, so that the CI has a chance to think of installing it?