@gasche 2022-01-07
Ah, I guess: the benchmark is not actually declared as a test (so it is not run by `runtest` and it is not obviously a test dependency); the build failure comes from the `@all` or `@check` targets that still try to build it anyway.