@avsm 2022-01-07
If it's a short running benchmark, then I'd prefer it to be declared as a test dependency and always run. (this sometimes finds crashes for us; the actual results aren't relevant in the test run obviously)