@mseri 2022-01-07
```
bench alias http/bench/runtest (exit 1)
(cd _build/default/http/bench && ./bench.exe)
Estimated testing time 40s (4 benchmarks x 10s). Change using '-quota'.
Uncaught exception:
  
  (Failure
   "conversion from int64 to int failed: -4253731984 is out of range")
```
Looks like core_bench does not like 32 bit systems