@mseri 2022-01-07
I have tried to run the benchmark on my old i5 a few times in fact.
Using `core_bench` I get the following:

```
Estimated testing time 40s (4 benchmarks x 10s). Change using '-quota'.

New comparison function:
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │  10.19ns │   4.00w │     17.04% │
│ Header.mem - 1 │   6.64ns │   4.00w │     11.11% │
│ Header.mem - 2 │  59.80ns │   4.00w │    100.00% │
│ Header.mem - 3 │  40.94ns │   4.00w │     68.46% │
└────────────────┴──────────┴─────────┴────────────┘

Old comparison function:
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │   9.95ns │   4.00w │     19.57% │
│ Header.mem - 1 │   6.99ns │   4.00w │     13.74% │
│ Header.mem - 2 │  50.86ns │   4.00w │    100.00% │
│ Header.mem - 3 │  33.17ns │   4.00w │     65.21% │
└────────────────┴──────────┴─────────┴────────────┘
```

From which it seems that for such short strings and my old computer the old implementation is faster.
I am curious to try on the laptop, but I will likely not be able before next week.