@mseri 2022-01-07
Tested on the macbook (core i7 from 2019, in battery mode).
```
Old comparison
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │  11.57ns │   4.00w │     21.70% │
│ Header.mem - 1 │   7.75ns │   4.00w │     14.53% │
│ Header.mem - 2 │  53.32ns │   4.00w │    100.00% │
│ Header.mem - 3 │  35.59ns │   4.00w │     66.75% │
└────────────────┴──────────┴─────────┴────────────┘


New comparison
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │  11.18ns │   4.00w │     21.98% │
│ Header.mem - 1 │   7.77ns │   4.00w │     15.28% │
│ Header.mem - 2 │  50.84ns │   4.00w │    100.00% │
│ Header.mem - 3 │  37.04ns │   4.00w │     72.85% │
└────────────────┴──────────┴─────────┴────────────┘
```