@mseri 2022-01-07
I have added one more check to the benchmark, of a difference that happens towards the end of the string.
Now one can see a large improvement
```
Old comparison
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │  11.73ns │   4.00w │     23.36% │
│ Header.mem - 1 │   7.61ns │   4.00w │     15.16% │
│ Header.mem - 2 │  50.22ns │   4.00w │    100.00% │
│ Header.mem - 3 │  35.62ns │   4.00w │     70.93% │
│ Header.mem - 4 │  48.70ns │   4.00w │     96.98% │
└────────────────┴──────────┴─────────┴────────────┘

New comparison
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │  11.58ns │   4.00w │     17.09% │
│ Header.mem - 1 │   7.78ns │   4.00w │     11.48% │
│ Header.mem - 2 │  67.79ns │   4.00w │    100.00% │
│ Header.mem - 3 │  46.45ns │   4.00w │     68.52% │
│ Header.mem - 4 │  21.38ns │   4.00w │     31.54% │
└────────────────┴──────────┴─────────┴────────────┘
```