@mseri 2022-01-07
There are (going to be) two more cases in the benchmark on the master branch, which contain differences only towards the end of the string. There you can see the improvement from the new PR

```
Old implementation
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │   7.47ns │   4.00w │     23.44% │
│ Header.mem - 1 │   4.87ns │   4.00w │     15.28% │
│ Header.mem - 2 │  31.56ns │   4.00w │     98.96% │
│ Header.mem - 3 │  22.23ns │   4.00w │     69.71% │
│ Header.mem - 4 │  30.85ns │   4.00w │     96.74% │
│ Header.mem - 5 │  31.89ns │   4.00w │    100.00% │
└────────────────┴──────────┴─────────┴────────────┘

New implementation
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ Header.mem - 0 │   7.34ns │   4.00w │     17.17% │
│ Header.mem - 1 │   4.97ns │   4.00w │     11.62% │
│ Header.mem - 2 │  42.74ns │   4.00w │    100.00% │
│ Header.mem - 3 │  29.68ns │   4.00w │     69.44% │
│ Header.mem - 4 │  39.90ns │   4.00w │     93.36% │
│ Header.mem - 5 │  13.44ns │   4.00w │     31.45% │
└────────────────┴──────────┴─────────┴────────────┘
```